+++ packages/common/package.json
@@ -21,8 +21,5 @@
     "multer": "^1.3.0",
     "reflect-metadata": "^0.1.12",
     "rxjs": "^6.0.0"
-  },
-  "devDependencies": {
-    "tsconfig-paths": "^3.6.0"
   }
 }
