+++ packages/core/package.json
@@ -16,7 +16,7 @@
     "class-transformer": "^0.1.9",
     "cli-color": "^1.2.0",
     "cors": "^2.8.4",
-    "deprecate": "^1.0.0",
+    "deprecate": "1.0.0",
     "express": "^4.16.3",
     "iterare": "^0.0.8",
     "multer": "^1.3.0",
