+++ packages/websockets/src/exceptions/base-ws-exception-filter.ts
