+++ packages/websockets/src/exceptions/index.ts
