+++ packages/websockets/src/exceptions/ws-exception.ts
