+++ packages/websockets/src/exceptions/ws-exceptions-handler.ts
