+++ packages/websockets/src/gateway-metadata-explorer.ts
