+++ packages/websockets/src/interfaces/gateway-metadata.interface.ts
