+++ packages/websockets/src/interfaces/gateway-middleware.interface.ts
