+++ packages/websockets/src/interfaces/index.ts
