+++ packages/websockets/src/interfaces/nest-gateway.interface.ts
