+++ packages/websockets/src/interfaces/observable-socket-server.interface.ts
