+++ packages/websockets/src/interfaces/on-gateway-connection.interface.ts
