+++ packages/websockets/src/interfaces/on-gateway-disconnect.interface.ts
