+++ packages/websockets/src/interfaces/on-gateway-init.interface.ts
