+++ packages/websockets/src/interfaces/web-socket-server.interface.ts
