+++ packages/websockets/src/interfaces/ws-response.interface.ts
