+++ packages/websockets/test/context/exception-filters.context.spec.ts
@@ -1,9 +1,9 @@
 import * as sinon from 'sinon';
 import { expect } from 'chai';
-import { UseFilters } from '../../../common/decorators/core/exception-filters.decorator';
-import { Catch } from '../../../common/decorators/core/catch.decorator';
-import { ExceptionFiltersContext } from '../../context/exception-filters-context';
-import { NestContainer } from '../../../core/injector/container';
+import { UseFilters } from '@nestjs/common/decorators/core/exception-filters.decorator';
+import { Catch } from '@nestjs/common/decorators/core/catch.decorator';
+import { ExceptionFiltersContext } from '@nestjs/websockets/context/exception-filters-context';
+import { NestContainer } from '@nestjs/core/injector/container';
 
 describe('ExceptionFiltersContext', () => {
   let moduleName: string;
