+++ packages/websockets/test/exceptions/ws-exception.spec.ts
@@ -1,6 +1,6 @@
 import * as sinon from 'sinon';
 import { expect } from 'chai';
-import { WsException } from '../../exceptions/ws-exception';
+import { WsException } from '@nestjs/websockets/exceptions/ws-exception';
 
 describe('WsException', () => {
   let instance: WsException;
