+++ packages/websockets/tsconfig.json
@@ -1,28 +1,17 @@
 {
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "module": "commonjs",
-    "declaration": true,
-    "noImplicitAny": false,
-    "noUnusedLocals": false,
-    "removeComments": false,
-    "noLib": false,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "target": "es2017",
-    "sourceMap": false,
-    "allowJs": false,
-    "baseUrl": "../../",
-    "strict": true,
-    "strictNullChecks": false,
-    "outDir": "../../node_modules/@nestjs/websockets"
+    "rootDir": "src",
+    "outDir": "lib"
   },
   "include": [
-    "*.ts",
-    "**/*.ts"
+    "src"
   ],
   "exclude": [
-    "node_modules/@types/reflect-metadata",
-    "node_modules/reflect-metadata",
-    "./**/*.spec.ts"
+    "lib"
+  ],
+  "references": [
+    { "path": "../common" },
+    { "path": "../core" }
   ]
 }
